+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - friend request rejected
     - friend removed
 - Added SonarQube analysis
+- Track alcohol-free beer separately in drink metadata and statistics.
 
 ## [1.2.1] - 2026-04-13
 ### Fixed